Изменил типы данных в генераторе для пчел

+++ app-main/src/main/java/ru/coderiders/bitemebee/service/BeeFamilyService.java
@@ -16,9 +16,9 @@ public interface BeeFamilyService {
 
     BeeFamilyRsDto update(@NonNull Long id, @NonNull BeeFamilyNoteRqDto beeFamilyNoteRqDto);
 
-    void updatePopulation(@NonNull Long id, @NonNull Integer dronePopulation,
-                          @NonNull Integer workerPopulation, @NonNull Integer queenPopulation,
-                          @NonNull Integer population);
+    void updatePopulation(@NonNull Long id, @NonNull Long dronePopulation,
+                          @NonNull Long workerPopulation, @NonNull Long queenPopulation,
+                          @NonNull Long population);
 
     void release(@NonNull Long id);
